+++ skeleton-tests/base.py
@@ -1,10 +1,15 @@
 # -*- coding: utf-8 -*-
 
+# from bobtemplates.plone.utils import safe_encode
 from bobtemplates.plone.utils import safe_unicode
 
 import contextlib
+import logging
 import os
 import subprocess
+
+
+logger = logging.getLogger("skeleton-tests")
 
 
 @contextlib.contextmanager
@@ -24,19 +29,20 @@ def file_exists(base_path, file_path):
 
 def run_skeleton_tox_env(wd, config):
     try:
-        test_result = subprocess.check_output(
-            ['tox', '-e', config.skeleton_tox_env],
+        returncode = subprocess.check_call(
+            ['tox', '-q', '-e', config.skeleton_tox_env],
             cwd=wd,
         )
-        print(u'\n{0}\n'.format(safe_unicode(test_result)))
+        # logger.debug(u'\n{0}\n'.format(safe_unicode(test_result)))
+        return returncode
     except subprocess.CalledProcessError as execinfo:
-        tox_msg = safe_unicode(
-            b''.join(bytes(execinfo.output)),
-        )
-        print(tox_msg)
-        tox_summary = safe_unicode(
-            b''.join(
-                execinfo.output.partition(b'__ summary __')[1:],
-            ),
-        )
-        assert execinfo.returncode == 0, '\n{0}'.format(tox_summary)
+        logger.debug(u'{0}'.format(safe_unicode(execinfo.output)))
+        return execinfo.returncode
+        # tox_msg = safe_unicode(execinfo.output)
+        # print(tox_msg)
+        # tox_summary = safe_unicode(
+        #     b''.join(
+        #         execinfo.output.partition(b'__ summary __')[1:],
+        #     ),
+        # )
+        # assert execinfo.returncode == 0  #, u'\n{0}'.format(tox_msg)